+++ src/main/kotlin/domain/lotto/LottoNumbers.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.lotto
 
 import java.util.SortedSet
