+++ src/main/kotlin/domain/money/Money.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.money
 
 data class Money(val value: Int) : Comparable<Money> {
     operator fun plus(other: Money): Money {
