+++ src/main/kotlin/domain/store/RandomLottoNumberGenerator.kt
@@ -1,4 +1,6 @@
-package domain
+package domain.store
+
+import domain.lotto.LottoNumbers
 
 interface RandomLottoNumberGenerator {
     fun generate(): LottoNumbers
