+++ src/test/kotlin/domain/statistics/WinningCategoryTest.kt
@@ -1,5 +1,6 @@
-package domain
+package domain.statistics
 
+import domain.money.Money
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
